add optional unexpected size filter
Adds a --check-size flag to grouping_series_stream_filter. If
set, will log a warning and skip files whenever a file in the
queue has a size that differs from the first file seen by that queue.
Intended use is to check for incomplete transfers in one particular
use case where this is a common failure mode.