adds scopt option parsing and sbt-assembly plugin
scopt provides reasonable command line option parsing, as
demonstrated in StatefulStats.scala. sbt-assembly is used to
package scopt and other dependency jars into a
thunder-streaming-assembly uber-jar, which if given to spark-submit
eliminates the need to specify additional dependencies explicitly
with the --jars flag. See comment in StatefulStats.scala for usage.