add wildcard support for input directory paths in feeder script
input directory paths in grouping_series_stream_feeder are now
matched by the glob module. The alphabetically last matching
directory will be selected and used. This allows for rerunning
the input script and automatically selecting the most recent
directory with a (sensible) timestamp in the directory name.

parameters in the nikita script are also updated; in particular,
time to wait after the last modified time before pushing the file
is increased from the default 1s to 5s, in hopes of smoothing out
some network file transfer issues.